Aggiunta la pagina Maps passaggio dei dati tra MainPage e Maps Implementare la pagina della mappa #2
