+++ react-app/src/scenes/graphsPage/index.jsx
